+++ packages/firebase_ml_vision/ios/Classes/FirebaseMlVisionPlugin.h
@@ -14,6 +14,12 @@
 @optional
 @end
 
+@protocol ModelManager
+@required
++ (void)modelName:(NSString *)modelName result:(FlutterResult)result;
+@optional
+@end
+
 @interface BarcodeDetector : NSObject <Detector>
 @end
 
@@ -24,4 +30,16 @@
 @end
 
 @interface TextRecognizer : NSObject <Detector>
+@end
+
+@interface LocalVisionEdgeDetector : NSObject <Detector>
+@end
+
+@interface RemoteVisionEdgeDetector : NSObject <Detector>
+@end
+
+@interface SetupLocalModel : NSObject <ModelManager>
+@end
+
+@interface SetupRemoteModel : NSObject <ModelManager>
 @end