+++ packages/firebase_ml_vision/lib/firebase_ml_vision.dart
@@ -18,3 +18,4 @@ part 'src/firebase_vision.dart';
 part 'src/image_labeler.dart';
 part 'src/text_recognizer.dart';
 part 'src/vision_edge.dart';
+part 'src/model_manager.dart';